Add the doc about Symfony <3.1
As the only way to be in this case is to run on Symfony 2.8, building a
feature in the bundle to do it ourselves is not worth it, especially given
that it would need to be configurable too anyway. It is simpler to ask the
users of the bundle to create an alias themselves (or to upgrade to
Symfony 3.1+).